double trouble
git-svn-id: http://svn.openzap.org/svn/openzap/trunk@329 a93c3328-9c30-0410-af19-c9cd2b2d52af

+++ libs/openzap/src/libteletone_detect.c
@@ -130,7 +130,7 @@ void teletone_goertzel_update(teletone_goertzel_state_t *goertzel_state,
 	}
 }
 
-#define teletone_goertzel_result(gs) (float)(((gs)->v3 * (gs)->v3 + (gs)->v2 * (gs)->v2 - (gs)->v2 * (gs)->v3 * (gs)->fac))
+#define teletone_goertzel_result(gs) (double)(((gs)->v3 * (gs)->v3 + (gs)->v2 * (gs)->v2 - (gs)->v2 * (gs)->v3 * (gs)->fac))
 
 void teletone_dtmf_detect_init (teletone_dtmf_detect_state_t *dtmf_detect_state, int sample_rate)
 {
